rgw: es_obj_metadata encode instance "null" for none-version obj
The instance of a none versioned object is empty.
In the bucket incremental sync of es zone, the instance is keep empty,
while in the full sync the instance is changed from empty to string "null".
As a result, the instance of the same none versioned object may be different,
especially when a bucket resynced by "radosgw-admin bucket sync init".

Signed-off-by: Shang Ding <dingshang@xsky.com>